% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{add_obs.prob_attr}
\alias{add_obs.prob_attr}
\title{Add attribute to each obs.prob}
\usage{
add_obs.prob_attr(obs.prob, type, subtype, fun)
}
\arguments{
\item{obs.prob}{the probability of observation matrix}

\item{type}{its type ("unbiased","trait", or "net")}

\item{subtype}{its subtype function used to make it}

\item{fun}{its conjugation function used to make it}
}
\value{
an obs.prob matrix with attributes to keep track of (sub)type and functions involved in its making
}
\description{
Add attribute to each obs.prob
}
\examples{
# Internal use
}
