% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.to.assoc.R
\name{adj.to.assoc}
\alias{adj.to.assoc}
\title{Adjacency matrix to association indices matrix}
\usage{
adj.to.assoc(Adj, effort = NULL, index = c("joint", "sri", "hw", "tw",
  "sqrt", "socaff", "both", "focal"), type = c("total", "yab"),
  mode = c("directed", "undirected", "max", "min", "upper", "lower",
  "plus"))
}
\arguments{
\item{Adj}{Adjacency matrix of edge weights}

\item{effort}{facultative vector of sampling effort per individual}

\item{index}{character string. Type of association index (cf. Whitehead, 2008).}

\item{type}{character string. Type of effort per individual provided. Either "yab", indicates the number of times an individual have been observed associated with no-one, or "total" for its total sampling effort (cf. Whitehead, 2008). Logically, the number or association of an individual cannot be smaller than the total number of observation for this individual.}

\item{mode}{}
}
\value{
A matrix of association index.
}
\description{
Calculate the association indices of choice from an adjacency matrix.
}
\examples{
set.seed(42)

n<- 5;nodes<- letters[1:n];
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[upper.tri(Adj)]<- sample(0:8,n*(n-1)/2,replace = TRUE)
Adj[lower.tri(Adj)]<- sample(0:8,n*(n-1)/2,replace = TRUE)
Adj

effort<- rowSums(Adj)+sample(5:8,n,replace = TRUE)

adj.to.assoc(Adj,effort,type="total",index="focal")

yab<- sample(5:8,n,replace = TRUE)

adj.to.assoc(Adj,yab,type = "yab",index="focal")
adj.to.assoc(Adj,yab,index="focal")
adj.to.assoc(Adj,yab,type = "total",index="focal")
}
