% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{adj_distance}
\alias{adj_distance}
\title{Wrapper to calculate adjacency/network distances}
\usage{
adj_distance(Bootstrap.list, method = c("group", "focal"), dist.fun)
}
\arguments{
\item{Bootstrap.list}{an output of Boot_scan()}

\item{method}{character scalar, indicate if the function should output the coefficient between the theoretical adjacency matrix and either the empirical one from group or focal method}

\item{dist.fun}{a matrix/network distance function taking an igraph object (or an adjacency matrix but transforms it to an igraph object first) and returns an ordered vector of vertex centralities}
}
\value{
a vector of coefficient of correlations between theoretical and empirical methods derived network node centralities.
}
\description{
Wrapper to calculate adjacency/network distances
}
\examples{
#Internal use in Simulation_script.R.
}
