% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{adjust.conditional.prob}
\alias{adjust.conditional.prob}
\title{Determine step-by-step conditional probabilities for non-zeroes scans
Internal use. Returns the CDF of the probability that the i-th dyad (with probability presence.prob[i]) is the first to yield a 1.}
\usage{
adjust.conditional.prob(presence.prob)
}
\arguments{
\item{presence.prob}{presence probability matrix (or vector)}
}
\value{
a cumulative distribution function of the probability that the i-th dyad (with probability presence.prob[i]) is the first to yield a 1
}
\description{
Determine step-by-step conditional probabilities for non-zeroes scans
Internal use. Returns the CDF of the probability that the i-th dyad (with probability presence.prob[i]) is the first to yield a 1.
}
\details{
Workflow is as follows: first simulate.zeroes.non.zeroes() determines which scans are all-zeroes and which are non-zeroes. Then for non zeroes, at a random order each dyad is drawn in order with conditional probability that: (1) there is at least one 1 in the scan, and (2) all the previous coins were zeroes. Once the first one is drawn, the rest are drawn with their regular probabilities. In details, cumulative density probability of each dyad (in a given random order) to be the first one to be a 1 is calculated, and a random draw determine which one is first, set the previous ones to zero, and draw the rest normally. cf. Supplmenentary material X.
}
\examples{
# Internal use.
}
