% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{association.index_CV}
\alias{association.index_CV}
\title{Coefficient of variation of association indices}
\usage{
association.index_CV(Adj, mode)
}
\arguments{
\item{Adj}{square integers matrix of occurences of dyads. WIP: implement method for association matrices...}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}
}
\value{
Coefficient of variation of association indices
}
\description{
Coefficient of variation of association indices
}
\examples{
# Internal use.
}
