% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot.group_scan.R
\name{Boot.group_scan}
\alias{Boot.group_scan}
\title{Bootstrap Adjacency
Perform a bootstrap of group scans with probabilities derived from a provided adjancecy matrix, to produce a new adjancecy matrix.}
\usage{
Boot.group_scan(
  Adj,
  total_scan,
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus"),
  output = c("list", "adjacency"),
  n.cores = (parallel::detectCores() - 1)
)
}
\arguments{
\item{Adj}{square integers matrix of occurences of dyads. WIP: implement method for association matrices...}

\item{total_scan}{integer, sampling effort. Note that 1/total_scan should be relatively small, increasingly small with increasing precision.}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}

\item{output}{Character scalar, specify if the function should return the list of scans, or reduce them into the bootstrapped adjacency matrix}

\item{n.cores}{number of threads to use while performingh the bootstrap}
}
\value{
according to output a list of scans, or the bootstrapped adjacency matrix
}
\description{
Bootstrap Adjacency
Perform a bootstrap of group scans with probabilities derived from a provided adjancecy matrix, to produce a new adjancecy matrix.
}
\examples{
set.seed(42)

n<- 5;nodes<- letters[1:n];
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:30,n*(n-1),replace = TRUE)
Adj

Boot.group_scan(Adj,42,output="adj")
Boots<- lapply(
  1:10,
    function(b){
        Boot.group_scan(Adj,42,mode = "directed",output = "adj",n.core=1)
    }
)
}
