% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Boot_calc.data}
\alias{Boot_calc.data}
\title{Calculate desired data from simulation
Internal use. To ease the recollection of a given bootstrap performed through Boot_scans() iterations alongside a parameters.list}
\usage{
Boot_calc.data(Bootstrap.list, method = c("group", "focal"))
}
\arguments{
\item{Bootstrap.list}{Bootstrap output object}

\item{method}{sampling method to retrieve data for.}
}
\value{
a data frame of desired data
}
\description{
Calculate desired data from simulation
Internal use. To ease the recollection of a given bootstrap performed through Boot_scans() iterations alongside a parameters.list
}
\examples{
#Internal use in Simulation_script.R.
}
