% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Boot_calc.data2}
\alias{Boot_calc.data2}
\title{WIP: revamping Boot_calc.data to be flexible in what it calculates}
\usage{
Boot_calc.data2(Bootstrap.list, fun.list, args.list = NULL, cname.list = NULL)
}
\arguments{
\item{cname.list}{}
}
\value{

}
\description{
WIP: revamping Boot_calc.data to be flexible in what it calculates
}
