% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Boot_get.param}
\alias{Boot_get.param}
\title{Retrieve specific simulation parameters of given Bootstrap
Internal use. To ease the recollection of a given bootstrap performed through Boot_scans() iterations alongside a parameters.list}
\usage{
Boot_get.param(parameters)
}
\arguments{
\item{parameters}{a list containing a combination of obs.prob, focal.list, and mode parameters. Element of parameters.list}
}
\value{
a data frame referencing the simulation parameters
}
\description{
Retrieve specific simulation parameters of given Bootstrap
Internal use. To ease the recollection of a given bootstrap performed through Boot_scans() iterations alongside a parameters.list
}
\examples{
#Internal use in Simulation_script.R.
}
