% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Boot_merge.from_Adj}
\alias{Boot_merge.from_Adj}
\title{Title}
\usage{
Boot_merge.from_Adj(
  Bootsrap.df,
  ADJ,
  net.fun.list,
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus"),
  cnames = NULL
)
}
\arguments{
\item{mode}{}
}
\value{

}
\description{
Title
}
