% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{boot_progress.param}
\alias{boot_progress.param}
\title{Bootstrap specific progress bar
Provide feedbacks on the simulation testing situation (which parameters, which combination of parameters over the whole list). Internal use.}
\usage{
boot_progress.param(p, parameters.list = parameters.list)
}
\arguments{
\item{p}{integer, index of the current combination of parameters within parameters.list.}

\item{parameters.list}{data frame of parameters for all the simulations}
}
\value{
display some feedbacks on the simulation testing situation (which parameters, which combination of parameters over the whole list).
}
\description{
Bootstrap specific progress bar
Provide feedbacks on the simulation testing situation (which parameters, which combination of parameters over the whole list). Internal use.
}
\examples{
#Internal use in Simulation_script.R.
}
