% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network _bootstrap_tools.R
\name{bootstrap_metric}
\alias{bootstrap_metric}
\title{TO WRITE}
\usage{
bootstrap_metric(scan.list, metric.fun, n.boot = 100)
}
\arguments{
\item{scan.list}{TO WRITE}

\item{n.boot}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
