% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.network.R
\name{build.network}
\alias{build.network}
\title{Network building function wrapper}
\usage{
build.network(Obs, id = "id", tar = "tar", boot = 10,
  rdraw = FALSE, replacement = TRUE, directed = FALSE, IDS = NULL)
}
\arguments{
\item{Obs}{WIP}

\item{id}{WIP}

\item{tar}{WIP}

\item{boot}{WIP}

\item{rdraw}{WIP}

\item{replacement}{WIP}

\item{directed}{WIP}

\item{IDS}{WIP}
}
\value{
WIP
}
\description{
Wrapper for building and bootstrapping networks
}
