% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{centrality_cor}
\alias{centrality_cor}
\title{Wrapper to calculate coefficient of correlation between centrality indices}
\usage{
centrality_cor(Bootstrap.list, method = c("group", "focal"), centrality.fun)
}
\arguments{
\item{Bootstrap.list}{an output of Boot_scan()}

\item{method}{character scalar, indicate if the function should output the coefficient between the theoretical adjacency matrix and either the empirical one from group or focal method}

\item{centrality.fun}{a centrality function taking an igraph object (or an adjacency matrix but transforms it to an igraph object first) and returns an ordered vector of vertex centralities}
}
\value{
a vector of coefficient of correlations between theoretical and empirical methods derived network node centralities.
}
\description{
Wrapper to calculate coefficient of correlation between centrality indices
}
\examples{
#Internal use in Simulation_script.R.
}
