% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfiBoot.R
\name{ConfiBoot}
\alias{ConfiBoot}
\title{Network Bootstrap of edge-lists}
\usage{
ConfiBoot(
  Obs,
  id = "id",
  tar = "tar",
  boot = 100,
  replacement = TRUE,
  proportion = 1,
  output = "list"
)
}
\arguments{
\item{Obs}{WIP}

\item{id}{WIP}

\item{tar}{WIP}

\item{boot}{WIP}

\item{replacement}{WIP}

\item{proportion}{WIP}

\item{output}{WIP}
}
\value{
WIP
}
\description{
Bootstrap edge lists where repeated lines represent dyads' weights
}
\examples{
print("WIP")
}
