% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.event.optimization_tools.R
\name{decide_use.rare.opti}
\alias{decide_use.rare.opti}
\title{Decide based on expected times if the otpimization for rare event should be used}
\usage{
decide_use.rare.opti(n, total_scan, max.obs = NULL)
}
\arguments{
\item{n}{number of node of the original network}

\item{total_scan}{sampling effort}

\item{max.obs}{maximum value of the weighted adjacency matrix of the original network}
}
\value{
a logical value meaning that the optimization for rare events should be use when TRUE is returned
}
\description{
Decide based on expected times if the otpimization for rare event should be used
}
\examples{
# Internal use
}
