% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.scan.R
\name{do.scan}
\alias{do.scan}
\title{Perform a scan
According to edge presence probability, or a reference adjacency matrix `Adj` coupled with a sample effort `total_scan`, perform a scan in the form of a theoretical binary adjacency matrix, and simulate the empirical obtention of a group scan and/or focal scan.}
\usage{
do.scan(
  Adj = NULL,
  total_scan = NULL,
  method = c("theoretical", "group", "focal", "both"),
  ...,
  check.defaults = TRUE
)
}
\arguments{
\item{Adj}{square integers matrix of occurences of dyads. Optional if using presence.prob. Update: now if presence prob is passed as Adj (thus all(Adj<1) is TRUE), it will be rightfully assigned to presence.prob. WIP: implement method for association matrices...}

\item{total_scan}{integer, sampling effort. Note that 1/total_scan should be relatively small, increasingly small with increasing precision. Optional if using presence.prob.}

\item{method}{Character scalar, specifies if the function should return a theoretical perfect group scan, an  empirical group scan (a similarly dimensioned matrix as Adj), or a focal scan (a vector representing the given focal's row in the group scan matrix).}

\item{...}{additional argument to be used, to use produce a scan in a desired way.}

\item{check.defaults}{logical. Should scan.default.args() be used? Implemented in case iterate_scan might provide everything neeed, but at the moment it seems buggy.}

\item{focal}{Only required for method = "focal" or "both" Character scalar, indicate which focal to consider for the scan.}

\item{presence.prob}{square probability matrix of presence (as in Bernouilli trial) of dyads. Optional if using Adj and total_scan.}

\item{obs.prob}{either :
\itemize{
 \item{"a dyad observation probability matrix (P.obs)"}{of same dimension as Adj}
 \item{"a dyad observation vector"}{subsetted similarly as Adj (through the non.diagonal() function for instance)}
 \item{"a systematic dyad observation (P.obs constant for all i,j)"}{should be in [0,1], assumed to be the case when only one value is inputed)}
}}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}

\item{Adj.subfun}{subsetting function of the adjacency matrix. Driven by igraph "mode" argument}
}
\value{
a list of the theoretical square binary matrix representing the whole group scan, and:
\itemize{
 \item{"method = theoretical"}{that's all,}
 \item{"method = group"}{the empirical group scan square binary matrix}
 \item{"method = focal"}{the empirical focal scan binary vector (and the focal identity as a "focal" attribute)}
 \item{"method = both"}{or the three above}
 }
}
\description{
Perform a scan
According to edge presence probability, or a reference adjacency matrix `Adj` coupled with a sample effort `total_scan`, perform a scan in the form of a theoretical binary adjacency matrix, and simulate the empirical obtention of a group scan and/or focal scan.
}
\examples{
set.seed(42)

n<- 5;nodes<- as.character(1:n);
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:42,n*(n-1),replace = TRUE)
Adj

presence.prob<- Binary.prob(Adj,50)
obs.prob<- matrix(runif(n*n,0,1),n,n);diag(obs.prob)<- 0

do.scan(Adj,50,method = "group",obs.prob = 0.9)
do.scan(Adj,50,method = "both",obs.prob = obs.prob)
do.scan(presence.prob,method = "focal")
do.scan(presence.prob,method = "focal",focal = "4")
do.scan(presence.prob,method = "focal",focal = 3)
}
