% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.to.adj.R
\name{edge.to.adj}
\alias{edge.to.adj}
\title{Convert an adjacency matrix to an edge list with repeated lines}
\usage{
edge.to.adj(edge, mode = c("directed", "undirected", "max", "min",
  "upper", "lower", "plus"))
}
\arguments{
\item{edge}{data.table or two-columns matrix. Each line represent an edge (order matters if the graph is directed), repetition of a given (ordered) pair represent the weight of an edge.}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}
}
\value{
adjacency matrix
}
\description{
Wrapper for igraph functions in order to obtain an adjacency matrix where the cell a pair represent the edge weight. Useful to obtain the desired association index after bootstrapping the weighted edge list.
}
\examples{
set.seed(42)

n<- 5;nodes<- letters[1:n];
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[upper.tri(Adj)]<- sample(0:2,n*(n-1)/2,replace = TRUE)
Adj[lower.tri(Adj)]<- sample(0:1,n*(n-1)/2,replace = TRUE)
Adj

edge<- ConfiNet::adj.to.edge(Adj,mode = "directed")
edge.to.adj(edge,mode="directed")
edge.up<- ConfiNet::adj.to.edge(Adj,mode = "upper")
edge.to.adj(edge.up,mode="upper")

edge.plus<- ConfiNet::adj.to.edge(Adj,mode = "plus")
edge.to.adj(edge.plus,mode="plus")

edge.low<- ConfiNet::adj.to.edge(Adj,mode = "lower")
edge.to.adj(edge.low,mode="lower")
}
