% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.window.R
\name{extend.window}
\alias{extend.window}
\title{Extend the time-window}
\usage{
extend.window(Obs, di, ds, min.obs, rdraw = FALSE, pctdraw = 1, mode = "dates")
}
\arguments{
\item{Obs}{data.table of dyadic observations.}

\item{di}{inferior date boundary of the desired time-window. Format should be "YYYY-MM-DD"}

\item{ds}{superior date boundary of the desired time-window. Format should be "YYYY-MM-DD"}

\item{min.obs}{minimum number (integer) of observations (rows) desired in the returned time-window, taking into consideration that the output might be randomly subsampled keeping only pctdraw \% of Obs.}

\item{rdraw}{logical. Indicates if the output will be randomly subsampled or not.}

\item{pctdraw}{if rdraw is TRUE, the proportion in (0,1) of Obs that will be kept in the later subsample}

\item{mode}{character string, either "dates" or "data.table", to choose the desirted output format}
}
\value{
Returns either the time-window as a subset data table of dyadic observation, or a vector of the two (closest) dates so that a time-window bounded by these will contain at least min.obs rows.
}
\description{
Extend the provided time-window to ensure that the extended one contains at least the minimum number of observation.
}
\examples{
set.seed(42);ID<- letters[1:10];

D.Obs<- round(runif(n = 40,min = 0,max = 15))
Obs<- data.table::data.table(date=as.Date(D.Obs-15,origin = "1991-03-30"),
                             id=sample(ID,40,replace = TRUE),
                             tar=sample(ID,40,replace = TRUE))[order(date)]

extend.window(Obs,"1991-03-20","1991-03-25",15)
extend.window(Obs,"1991-03-20","1991-03-25",20,rdraw = TRUE,pctdraw = .90,mode="data.table")
}
