% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Frobenius_from_adjacency}
\alias{Frobenius_from_adjacency}
\title{Simple Frobenius distance calculation}
\usage{
Frobenius_from_adjacency(X, Y = 0, ...)
}
\arguments{
\item{X}{an adjacency matrix}

\item{Y}{another adjacency matrix, otherwise the distance measuredd is ||X||}

\item{...}{for compatibility when mode has to be inputted}
}
\value{
the Frobenius distance
}
\description{
Simple Frobenius distance calculation
}
\examples{
#Internal use in Simulation_script.R.
}
