% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Get.data}
\alias{Get.data}
\title{Retrieve data from list of network bootstrap results
To be used in a loop/lapply. format them in a ready to be analyzed data frame.}
\usage{
Get.data(B, Bootstrap.list)
}
\arguments{
\item{B}{index of the network from which data are collected.}

\item{Bootstrap.list}{list of bootstrap of a given network, through different parameters. Now should receive an unscaled bootstraplist that will also be post scaled.}
}
\value{
a data frame of collected data per network and per parameter combination
}
\description{
Retrieve data from list of network bootstrap results
To be used in a loop/lapply. format them in a ready to be analyzed data frame.
}
\examples{
#Internal use in Simulation_script.R.
}
