% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SN_metric_tools.R
\name{GlobalCC}
\alias{GlobalCC}
\title{TO WRITE}
\usage{
GlobalCC(Adj, mode)
}
\arguments{
\item{Adj}{TO WRITE}

\item{mode}{TO WRITE}
}
\value{
TO WRITE
}
\description{
TO WRITE
}
\examples{
# TO WRITE
}
