% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{initialize_parameter_list}
\alias{initialize_parameter_list}
\title{Create a list of parameter combinations
to be used in simulations of network sampling}
\usage{
initialize_parameter_list(ADJ, obs.bias_list, focal.bias_list)
}
\arguments{
\item{ADJ}{list of networks' adjacency matrices}

\item{obs.bias_list}{list of lists containing each a bias.fun_list, bias.subtype_list, and type elements, to be processed internally by `make_obs.prob.bias_list`}

\item{focal.bias_list}{a (flat) list of functions of (n,Adj) to be passed as focal.list argument of Bootscan (can be also "even" as the default special case).}
}
\value{
a list of lists containing: (1) a matrix `obs.prob` created for each matrix of `ADJ` from functions in `obs.bias_list`, and (2) a `focal.list` being either a function or "even". Both elements have differents attributes to keep track of (sub)type of bias.
}
\description{
Create a list of parameter combinations
to be used in simulations of network sampling
}
\examples{
#Internal use in Simulation_script.R.
}
