% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{initialize_parameters}
\alias{initialize_parameters}
\title{Create a list of parameter combinations
to be used in simulations of network sampling}
\usage{
initialize_parameters(
  Adj,
  total_scan,
  n.cores = (parallel::detectCores() - 1),
  cl = NULL
)
}
\arguments{
\item{Adj}{reference adjacency matrix (especially for network-based observation bias)}

\item{total_scan}{integer, total number of scan to be performed in each simulation}

\item{n.cores}{number of cores to use to generate (in parallel) the parameter combinations list}

\item{cl}{optional cluster object to use, otherwise automatically created when n.cores is inputted}
}
\value{
a list of lists of the different parameter combinations.
}
\description{
Create a list of parameter combinations
to be used in simulations of network sampling
}
\examples{
#Internal use in Simulation_script.R.
}
