% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KenBoot.R
\name{KenBoot}
\alias{KenBoot}
\title{Network Bootstrap of edge-lists}
\usage{
KenBoot(Adj, effort, type = c("total", "yab"), boot = 100,
  replacement = TRUE, proportion = 1, index = c("joint", "sri", "hw",
  "tw", "sqrt", "socaff", "both", "focal"), mode = c("directed",
  "undirected", "max", "min", "upper", "lower", "plus"),
  output = c("adjacency", "edge", "assoc", "graph", "metric"),
  FUN = NULL, ...)
}
\arguments{
\item{Adj}{raw adjacency matrix}

\item{effort}{vector of sampling effort per individual. WIP}

\item{type}{type of sampling effort. cf \link[KenNet]{adj.to.assoc}}

\item{boot}{number of bootstrap iteration. WIP}

\item{replacement}{resampling with replacement or not. WIP}

\item{proportion}{proportion to sub-sample, within [0,1[. WIP}

\item{index}{Type of desired association index. At the molment only focal method implemented. WIP}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}

\item{output}{desired output of the bootstrap: edge lists, adjacency matrices, association index matrices, igraph graphs, or network metrics}

\item{FUN}{if output is "metric", indicate which function to use to calculate said metric.}

\item{...}{if output is "metric", arguments for function FUN.}
}
\value{
according to output, returns a list of edge lists, adjacency matrices, association index matrices, or igraph graphs
}
\description{
Bootstrap edge lists where repeated lines represent dyads' weights
}
\examples{
print("WIP")
}
