% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Laplacian_spectral.dist}
\alias{Laplacian_spectral.dist}
\title{Graph Laplacian spectral distance}
\usage{
Laplacian_spectral.dist(X, Y = 0, ...)
}
\arguments{
\item{X}{an adjacency matrix}

\item{Y}{another adjacency matrix, otherwise the distance measuredd is ||X||}

\item{...}{for compatibility when mode has to be inputted}
}
\value{
the Laplacian spectral distance
}
\description{
Graph Laplacian spectral distance
}
\examples{
#Internal use for now only.
}
