% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{make_global_obs.prob}
\alias{make_global_obs.prob}
\title{Gather all obs.prob list from a list of lists}
\usage{
make_global_obs.prob(global_list)
}
\arguments{
\item{global_list}{list of lists containing each a bias.fun_list, bias.subtype_list, and type elements, to be processed internally by `make_obs.prob.bias_list`}
}
\value{
a list of elements containing: `obs.prob_fun` as the function to use to build the matrix obs.prob, and attributes to keep track of (sub)type and functions involved in its making
}
\description{
Gather all obs.prob list from a list of lists
}
\examples{
# Internal use
}
