% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{make_obs.prob.bias_list}
\alias{make_obs.prob.bias_list}
\title{Make a list of obs.prob matrices given list of functions to be composed}
\usage{
make_obs.prob.bias_list(
  bias.fun_list,
  bias.subtype_list,
  type = c("unbiased", "trait", "net")
)
}
\arguments{
\item{bias.fun_list}{function of (i,j,Adj) to conjugate two individual values (e.g. function(i,j,Adj) {bias.fun(bias.subtype(i,Adj),bias.subtype(j,Adj))})}

\item{bias.subtype_list}{function of (i,Adj) to get an individual value (can be trait- or net-based, or unbiased)(e.g. function(i,j,Adj) {bias.fun(bias.subtype(i,Adj),bias.subtype(j,Adj))})}

\item{type}{character, either "unbiased" (`bias.subtype_list` passed as a [0,1] numeric), "trait" for trait-based bias (function(i,j,Adj) {bias.fun(bias.subtype(i,Adj),bias.subtype(j,Adj))}), or "net" for net-based bias (centrality<- bias.subtype(Adj);bias.fun(centrality[i],centrality[j])).}
}
\value{
a list of elements containing: `obs.prob_fun` as the function to use to build the matrix obs.prob, and attributes to keep track of (sub)type and functions involved in its making
}
\description{
Make a list of obs.prob matrices given list of functions to be composed
}
\examples{
# Internal use
}
