% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.boot.R
\name{metric.boot}
\alias{metric.boot}
\title{Calculate network metric}
\usage{
metric.boot(graph, FUN, ...)
}
\arguments{
\item{graph}{graph (igraph) for which the metric will be calculated}

\item{FUN}{function to calculate the desired metric}

\item{...}{arguments of FUN}
}
\value{
output of FUN
}
\description{
Wrapper for internal use.
}
\examples{
print("internal use. Look around.")
}
