% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.prob_tools.R
\name{n.observed_edges}
\alias{n.observed_edges}
\title{Get number of edge observations (for group scans with unobservable individuals)
quantify actual edge-wise sampling effort, considering that some weren't observable in all group scans.Internal use.}
\usage{
n.observed_edges(
  scan_list,
  diag = 0,
  use.rare.opti = FALSE,
  obs.prob = NULL,
  n.zeros = NULL
)
}
\arguments{
\item{scan_list}{list of binary group scans, with NAs when the dyad was not observable.}

\item{diag}{integer (mostly), value to replace the diagonal of the output matrix with. Use NULL if you consider self-loop (untested).}

\item{use.rare.opti}{logical: should the optimization for rare event be used?}

\item{obs.prob}{either :
\itemize{
 \item{"a dyad observation probability matrix (P.obs)"}{of same dimension as Adj}
 \item{"a dyad observation vector"}{subsetted similarly as Adj (through the non.diagonal() function for instance)}
 \item{"a systematic dyad observation (P.obs constant for all i,j)"}{should be in [0,1], assumed to be the case when only one value is inputed)}
}}

\item{n.zeros}{integer, the attribute outputed by `simulate_zeros.non.zeros`, representing the number of full-zero scans. Used only when use.rare.opti=TRUE}
}
\value{
a square matrix with element quantifying how many time a dyad has been sampled
}
\description{
Get number of edge observations (for group scans with unobservable individuals)
quantify actual edge-wise sampling effort, considering that some weren't observable in all group scans.Internal use.
}
\examples{
#internal use.
}
