% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{net.metric.diff}
\alias{net.metric.diff}
\title{Wrapper to calculate difference between whole-group network metric}
\usage{
net.metric.diff(Bootstrap.list, method = c("group", "focal"), network.fun)
}
\arguments{
\item{Bootstrap.list}{an output of Boot_scan()}

\item{method}{character scalar, indicate if the function should output the coefficient between the theoretical adjacency matrix and either the empirical one from group or focal method}

\item{network.fun}{a whole-group network metric function taking an igraph object (or an adjacency matrix but transforms it to an igraph object first) and returns a metric}
}
\value{
the difference between metrics of networks obtained via theoretical and empirical methods.
}
\description{
Wrapper to calculate difference between whole-group network metric
}
\examples{
#Internal use in Simulation_script.R.
}
