% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.to.assoc.R
\name{non.zero.non.diag}
\alias{non.zero.non.diag}
\title{Identify coordinates of non null non diagonal elements of a matrix}
\usage{
non.zero.non.diag(M)
}
\arguments{
\item{M}{a square matrix}
}
\value{
an array of row and col indices of the positive and non diagonal elements
}
\description{
Identify coordinates of non null non diagonal elements of a matrix
}
\examples{
M<- matrix(sample(1:10,16,replace = TRUE),4,4)
diagonal<- function(M) {upper.tri(M,diag = TRUE)&!upper.tri(M,diag = FALSE)}

KenNet:::non.zero.non.diag(M)

}
