% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observable_edges.R
\name{observable_edges}
\alias{observable_edges}
\title{Hide unobservable edges
Simulate that some dyads might reasonable not be observable during a group scan}
\usage{
observable_edges(Scan, obs.prob = NULL, keep = FALSE)
}
\arguments{
\item{Scan}{a square matrix}

\item{obs.prob}{either :
\itemize{
 \item{"a dyad observation obs.probability matrix"}{of same dimension as Scan}
 \item{"a dyad observation vector"}{subsetted similarly as Scan (through the non.diagonal() function for instance)}
 \item{"a general dyad observation obs.probability"}{should be in [0,1], assumed to be the case when only one value is inputed)}
}}

\item{keep}{logical. Relevant if group scans are performed. Indicate if the original "theoretical" group scan should be kept track of.}
}
\value{
a similar matrix as Scan where some non diagonal edges have a obs.probability to be NAs. If keep is TRUE, returns a list with theoretical and observed scan.
}
\description{
Hide unobservable edges
Simulate that some dyads might reasonable not be observable during a group scan
}
\examples{
set.seed(42)

n<- 6;
Scan<- matrix(sample(c(1,0),n*n,replace=TRUE),n,n);diag(Scan)<- 0
obs.prob<- matrix(runif(n*n,0,1),n,n);diag(obs.prob)<- 0
traits<- rnorm(nrow(Scan),0,1)
trait.bias_fun<- function(x) {traits[x]}
obs.prob.trait<- obs.prob_bias(Scan,sum,bias_fun = trait.bias_fun)
obs.prob.single<- runif(1,0,1)

observable_edges(Scan,obs.prob)
observable_edges(Scan,obs.prob.trait)
observable_edges(Scan,obs.prob.single)
}
