% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{opti.expected.time}
\alias{opti.expected.time}
\title{Placeholder for calculating expected time using the optimization for rare event}
\usage{
opti.expected.time(n, total_scan, max.obs)
}
\arguments{
\item{n}{number of node of the original network}

\item{total_scan}{sampling effort}

\item{max.obs}{maximum value of the weighted adjacency matrix of the original network}
}
\value{
a time value to be compared with the one using the standard approach
}
\description{
Placeholder for calculating expected time using the optimization for rare event
}
\examples{
# Internal use in decide_use.rare.opti
}
