% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.tools.R
\name{reverse_i.n}
\alias{reverse_i.n}
\title{"Reverse" matrix rows and column order
For a matrix M with dim(M)=(n,m), swap element M[i,j] swapped by M[n-i+1,m-j+1]}
\usage{
reverse_i.n(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
a similarly dimensioned matrix with element M[i,j] swapped by M[n-i+1,m-j+1]
}
\description{
"Reverse" matrix rows and column order
For a matrix M with dim(M)=(n,m), swap element M[i,j] swapped by M[n-i+1,m-j+1]
}
\examples{
reverse_i.n(matrix(runif(15),3,5))
}
