% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.focal.R
\name{simu.focal}
\alias{simu.focal}
\title{Simulate a focal scan}
\usage{
simu.focal(
  n_focals,
  n_scans,
  n_nodes,
  nodes_names = as.character(1:n_nodes),
  focals = nodes_names,
  mode = c("directed", "undirected", "plus"),
  output = c("scans", "egos", "adjacency")
)
}
\arguments{
\item{n_focals}{integer, number of focals to simulate.}

\item{n_scans}{integer, number of scans to simulate.}

\item{n_nodes}{integer, number of nodes to simulate.}

\item{nodes_names}{optional character vector of the nodes' names, otherwise simply numbered.}

\item{focals}{optional character vector of the nodes' names, otherwise simply numbered.}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}

\item{output}{Character scalar, specifies if the function should return a list of lists of scans (n_focals-list of n_scans vectors), a list of egocentric network (n_focals-list, can have several per node), or an adjacency matrix. Only adjacency matrices can be undirected (???).}
}
\value{
depends on output
}
\description{
Simulate a list of instantaneous adjacency matrices
}
\examples{
set.seed(42)

simu.focal(n_focals = 15,n_scans = 5,n_nodes = 10,mode = "directed",output = "scan")
simu.focal(n_focals = 15,n_scans = 5,n_nodes = 10,mode = "directed",output = "ego")
simu.focal(n_focals = 15,n_scans = 5,n_nodes = 10,mode = "directed",output = "adj")
simu.focal(n_focals = 15,n_scans = 5,n_nodes = 10,mode = "plus",output = "adj")
}
