% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.scan.R
\name{simu.scan}
\alias{simu.scan}
\title{Simulate a group scan}
\usage{
simu.scan(
  n_scans,
  n_nodes,
  nodes_names = as.character(1:n_nodes),
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus")
)
}
\arguments{
\item{n_scans}{integer, number of scans to simulate.}

\item{n_nodes}{integer, number of nodes to simulate.}

\item{nodes_names}{optional character vector of the nodes' names, otherwise simply numbered.}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}
}
\value{
list of adjacency matrices
}
\description{
Simulate a list of instantaneous adjacency matrices
}
\examples{
set.seed(42)
n<- 10L;nodes<- letters[1:n];n_scans<- 50L;

simu.scan(n_scans,n_nodes = n,mode = "undirected")
simu.scan(n_scans,n_nodes = n,mode = "directed")
simu.scan(n_scans,n_nodes = n,nodes_names = nodes,mode = "plus")

}
