% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{simulate.zeroes.non.zeroes}
\alias{simulate.zeroes.non.zeroes}
\title{Simulate which scan returns an all-zeroes matrix}
\usage{
\method{simulate}{zeroes.non.zeroes}(total_scan, presence.prob)
}
\arguments{
\item{total_scan}{integer, sampling effort}

\item{presence.prob}{presence probability matrix (or vector)}
}
\value{
a list of zeroes (all-zeroes scans) and NULL (non-zeroes scans to be later performed)
}
\description{
Simulate which scan returns an all-zeroes matrix
}
\examples{
# Internal use
}
