% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.event.optimization_tools.R
\name{simulate_zeros.non.zeros.old}
\alias{simulate_zeros.non.zeros.old}
\title{Simulate which scan returns an all-zeros matrix
(old: will probably be deprecated because inefficient)}
\usage{
simulate_zeros.non.zeros.old(total_scan, presence.prob, method)
}
\arguments{
\item{total_scan}{integer, sampling effort}

\item{presence.prob}{presence probability matrix (or vector)}

\item{method}{Character scalar, specifies if the function should return a theoretical perfect group scan, an  empirical group scan (a similarly dimensioned matrix as Adj), or a focal scan (a vector representing the given focal's row in the group scan matrix).}
}
\value{
a list of zero-matrices (all-zeros scans) and NULL (non-zeros scans to be later performed)
}
\description{
Simulate which scan returns an all-zeros matrix
(old: will probably be deprecated because inefficient)
}
\examples{
# Internal use
}
