% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.event.optimization_tools.R
\name{standard.expected.time}
\alias{standard.expected.time}
\title{Placeholder for calculating expected time using the standard scan method
i.e. performing all scans}
\usage{
standard.expected.time(n, total_scan, max.obs)
}
\arguments{
\item{n}{number of node of the original network}

\item{total_scan}{sampling effort}

\item{max.obs}{maximum value of the weighted adjacency matrix of the original network}
}
\value{
a time value to be compared with the one using the optimization for rare event
}
\description{
Placeholder for calculating expected time using the standard scan method
i.e. performing all scans
}
\examples{
# Internal use in decide_use.rare.opti
}
