% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_up.scans.R
\name{sum_up.scan}
\alias{sum_up.scan}
\title{Sum up and flatten list of binary scans
Internal use. Sum up binary adjacency matrices or focal vectors according to the method used. sum_up.scan*s*() is a wrapper for sum_up.scan() that returns summed_up scans for the differnet requested methods0}
\usage{
sum_up.scan(
  scan_list,
  method,
  mode = NULL,
  scaled,
  use.rare.opti = FALSE,
  obs.prob = NULL
)
}
\arguments{
\item{scan_list}{list of binary adjacency matrices, binary focal vectors, or both, outputed by iterating do.scan()}

\item{method}{Character scalar, specify if the function should use a whole group or a focal scan sampling method (or both).}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. Default here is directed. Possible values are: directed, undirected, upper, lower, max, min, plus. Added vector too. See details \link[igraph]{graph_from_adjacency_matrix}.}

\item{scaled}{logical, specifies if adjacency data should be scaled by sampling effort.}
}
\value{
An integer adjacency matrix, or a list of two if method = "both". Now returns NAs for dyad scaled with no observation at all.
}
\description{
Sum up and flatten list of binary scans
Internal use. Sum up binary adjacency matrices or focal vectors according to the method used. sum_up.scan*s*() is a wrapper for sum_up.scan() that returns summed_up scans for the differnet requested methods0
}
\examples{
#Internal use for readability
}
