% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_up.scans.R
\name{sum_up.scans}
\alias{sum_up.scans}
\title{Sum up and flatten list of binary scans
Internal use. Sum up binary adjacency matrices or focal vectors according to the method used}
\usage{
sum_up.scans(
  Adj,
  scan_list,
  scaled = FALSE,
  method = c("group", "focal", "both")
)
}
\arguments{
\item{Adj}{square integers matrix of occurences of dyads. WIP: implement method for association matrices...}

\item{scan_list}{list of binary adjacency matrices, binary focal vectors, or both, outputed by iterating do.scan()}

\item{scaled}{logical, specifies if adjacency data should be scaled by sampling effort.}

\item{method}{Character scalar, specify if the function should use a whole group or a focal scan sampling method (or both).}
}
\value{
A non-binary adjacency matrix, or a list of two if method = "both"
}
\description{
Sum up and flatten list of binary scans
Internal use. Sum up binary adjacency matrices or focal vectors according to the method used
}
\examples{
#Internal use for readability
}
