% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{weighted.clustering.coeff}
\alias{weighted.clustering.coeff}
\title{Weighted clustering coefficient}
\usage{
weighted.clustering.coeff(graph, mode)
}
\arguments{
\item{graph}{a graph}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}
}
\value{
the clustering coefficient
}
\description{
Weighted clustering coefficient
}
\examples{
# Internal use.
}
