% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{adjacency_cor}
\alias{adjacency_cor}
\title{Calculate the correlation coefficient between "flattened" adjacency matrices
Provide the correlation coefficient between the theoretical adjacency matrix and either the empirical one from group or focal method}
\usage{
adjacency_cor(
  Bootstrap.list,
  what = c("observed", "focal"),
  n.boot = length(Bootstrap.list)
)
}
\arguments{
\item{Bootstrap.list}{an output of Boot_scan()}

\item{what}{character scalar, indicate if the function should output the coefficient between the theoretical adjacency matrix and either the empirical one from group or focal method}

\item{n.boot}{number of Bootstrap to go through}
}
\value{
a vector of correlation coefficients
}
\description{
Calculate the correlation coefficient between "flattened" adjacency matrices
Provide the correlation coefficient between the theoretical adjacency matrix and either the empirical one from group or focal method
}
\examples{
#Internal use in Simulation_script.R.
}
