% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.scan.R
\name{do.scan}
\alias{do.scan}
\title{Single group scan
Generate a random whole group or focal scan based on provided reference adjacency matrix and sampling effort.}
\usage{
do.scan(
  Adj,
  total_scan,
  focal = NULL,
  obs.prob = NULL,
  keep = FALSE,
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus"),
  method = c("group", "focal", "both"),
  Adj.subfun = NULL,
  prob = NULL
)
}
\arguments{
\item{Adj}{square integers matrix of occurences of dyads. WIP: implement method for association matrices...}

\item{total_scan}{integer, sampling effort. Note that 1/total_scan should be relatively small, increasingly small with increasing precision.}

\item{focal}{Character scalar, indicate which focal to consider for the scan.}

\item{obs.prob}{either :
\itemize{
 \item{"a dyad observation obs.probability matrix"}{of same dimension as Adj}
 \item{"a dyad observation vector"}{subsetted similarly as Adj (through the non.diagonal() function for instance)}
 \item{"a general dyad observation obs.probability"}{should be in [0,1], assumed to be the case when only one value is inputed)}
}}

\item{keep}{logical. Relevant if group scans are performed. Indicate if the original "theoretical" group scan should be kept track of.}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}

\item{method}{Character scalar, specifies if the function should return a whole-group scan (a similarly dimensioned matrix as Adj), or a focal scan (a vector representing the given focal's row in the group scan matrix).}

\item{Adj.subfun}{subsetting function of the adjacency matrix. Driven by igraph "mode" argument}

\item{prob}{output of Binary.prob function. Internal use when wrapped with Boot_scans, to speed up bootstrap process}
}
\value{
a square binary matrix representing the whle group scan
}
\description{
Single group scan
Generate a random whole group or focal scan based on provided reference adjacency matrix and sampling effort.
}
\examples{
set.seed(42)

n<- 5;nodes<- letters[1:n];
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- sample(0:30,n*(n-1),replace = TRUE)
Adj

do.scan(Adj,42)
do.scan(Adj,42,"c",method = "focal")
Reduce("+",lapply(1:42,function(s) do.scan(Adj,42)))

focal.list<- sample(nodes,42,replace = TRUE)
table(focal.list)
L<- lapply(1:42,function(s) do.scan(Adj,42,focal.list[s],method = "both"))

list(group = round(Reduce("+",lapply(L,function(l) l$group))/42,3),
     focal = {foc<- do.call(rbind,
                            lapply(nodes,
                                   function(node) {
                                     L.focal<- L[sapply(L,function(l) attr(l$focal,"focal"))==node]
                                     round(
                                           Reduce("+",
                                                  lapply(L.focal,
                                                         function(l) {
                                                           l$focal
                                                         }
                                                  )
                                           )/length(L.focal),3
                                     )
                                   }
                             )
                     );
              row.names(foc)<-nodes
                             foc}
)
}
