% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_tools.R
\name{Frobenius_from_adjacency}
\alias{Frobenius_from_adjacency}
\title{Simple Frobenius distance calculation}
\usage{
Frobenius_from_adjacency(X, Y = 0, mode)
}
\arguments{
\item{X}{an adjacency matrix}

\item{Y}{another adjacency matrix, otherwise the distance measuredd is ||X||}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. Default here is directed. Possible values are: directed, undirected, upper, lower, max, min, plus. Added vector too. See details \link[igraph]{graph_from_adjacency_matrix}.}

\item{...}{for compatibility when mode has to be inputted}
}
\value{
the Frobenius distance

the Frobenius distance
}
\description{
Simple Frobenius distance calculation
}
\examples{
#Internal use in Simulation_script.R.
Frobenius distance calculation
considers if the adjacency matrix is symetric or not (and avoid redundant information)

}
