% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.list.R
\name{make_focal.list}
\alias{make_focal.list}
\title{Produce focal.list}
\usage{
make_focal.list(Adj, total_scan, focal.prob_fun = NULL)
}
\arguments{
\item{Adj}{square integers matrix of occurences of dyads.}

\item{total_scan}{integer, sampling effort. Note that 1/total_scan should be relatively small, increasingly small with increasing precision. Optional if using presence.prob.}

\item{focal.prob_fun}{a user-defined function of (n) that output a probability of being focal for each node}
}
\value{
a vector of focals (as integers)
}
\description{
Produce focal.list
}
\examples{
set.seed(42)
n<- 6;nodes<- as.character(1:n);
total_scan<- 20;n.boot<- 5;

Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[non.diagonal(Adj)]<- round(runif(n*(n-1),0,total_scan*.50))

make_focal.list(Adj,total_scan)
make_focal.list(Adj,total_scan,focal.prob_fun = function(n) 1:n)
make_focal.list(Adj,total_scan,focal.prob_fun = function(n) compute.strength(Adj,"directed"))
}
