% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.boot.R
\name{metric.boot}
\alias{metric.boot}
\title{Calculate network metric}
\usage{
metric.boot(graph, FUN, args.list)
}
\arguments{
\item{graph}{graph (igraph) or list of graphsp for which the metric will be calculated}

\item{FUN}{function to calculate the desired metric. Should take at least}

\item{args.list}{arguments of FUN. graph will be internally used as a mandatory argument (hopefully being called graph is consistent throughout functions...)}
}
\value{
output of FUN, or list of outputs
}
\description{
Wrapper for internal use.
}
\examples{
# set.seed(42)
n<- 10;nodes<- letters[1:n];
Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
Adj[upper.tri(Adj)]<- sample(0:8,n*(n-1)/2,replace = TRUE)
Adj[lower.tri(Adj)]<- sample(0:8,n*(n-1)/2,replace = TRUE)
Adj

effort<- rowSums(Adj)+sample(5:8,n,replace = TRUE)
effort

boot<- KenNet::KenBoot(Adj,effort,10,replacement = TRUE,proportion = 1.0,
        index = "focal",type = "total",mode = "plus",output = "graph")
metric.boot(graph = boot[[2]],igraph::transitivity,args.list = list(type="global"))
metric.boot(boot,igraph::strength,args.list = list(mode="total",loops=FALSE))
}
