% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.tools.R
\name{n.observed_edges}
\alias{n.observed_edges}
\title{Get number of edge observations (for group scans with unobservable individuals)
quantify actual edge-wise sampling effort, considering that some weren't observable in all group scans.Internal use.}
\usage{
n.observed_edges(scan_list, diag = 0)
}
\arguments{
\item{scan_list}{list of binary group scans, with NAs when the dyad was not observable.}

\item{diag}{integer (mostly), value to replace the diagonal of the output matrix with. Use NULL if you consider self-loop (untested).}
}
\value{
a square matrix with element quantifying how many time a dyad has been sampled
}
\description{
Get number of edge observations (for group scans with unobservable individuals)
quantify actual edge-wise sampling effort, considering that some weren't observable in all group scans.Internal use.
}
\examples{
#internal use.
}
