% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_up.scans.R
\name{sum_up.scans}
\alias{sum_up.scans}
\title{Sum up and flatten list of binary scans
Internal use. Sum up binary adjacency matrices or focal vectors according to the method used}
\usage{
sum_up.scans(
  Adj,
  scan_list,
  scaled = FALSE,
  keep = FALSE,
  method = c("group", "focal", "both"),
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus")
)
}
\arguments{
\item{Adj}{square integers matrix of occurences of dyads. WIP: implement method for association matrices...}

\item{scan_list}{list of binary adjacency matrices, binary focal vectors, or both, outputed by iterating do.scan()}

\item{scaled}{logical, specifies if adjacency data should be scaled by sampling effort.}

\item{keep}{logical. Relevant if group scans are performed. Indicate if the original "theoretical" group scan should be kept track of.}

\item{method}{Character scalar, specify if the function should use a whole group or a focal scan sampling method (or both).}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.}
}
\value{
A non-binary adjacency matrix, or a list of two if method = "both"
}
\description{
Sum up and flatten list of binary scans
Internal use. Sum up binary adjacency matrices or focal vectors according to the method used
}
\examples{
#Internal use for readability
}
